LiveProjects/DataScrape/DataScrape_3-25/requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
idna==2.8
numpy==1.16.2
pytz==2018.9
requests==2.21.0
selenium==3.141.0
soupsieve==1.7.2
urllib3==1.24.1
lxml==4.3.3
nasa-api-wrapper==0.1.5
html5lib==1.0.1
pyqt5==5.12.1
w3lib==1.20.0
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
google-auth==1.6.3
python-launch-library==0.6
python-google-calendar-api==0.5